Higher Order Component to do Badger logic.  Easily create custom Badger Buttons
