+++ src/hoc/BadgerBase/index.js
